requirements.txt
pandas
numpy
matplotlib
scikit-learn
scikit-learn-extra
joblib
streamlit
plotly
umap-learn
hdbscan
groq